requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
flask-restful
flask-marshmallow
flask-socketio
marshmallow-sqlalchemy
python-dotenv
flask-mail
celery
redis
uvicorn
asgiref
pg8000
asgrief